split files based on root index js file
