+++ src/components/NameInput/NameInput.scss
@@ -1,0 +1,6 @@
+.NameInput {
+  font-size: 24px;
+  border: 0px;
+  border-bottom: 2px solid black;
+  font-weight: bold;
+}
