+++ src/components/PokeImage/PokeImage.scss
@@ -1,0 +1,3 @@
+.PokeImage {
+  float:left
+}
